Fixing login functionality - Authentication needed to be setup on https://console.firebase.google.com/u/0/project/newtritionist-app-angular/authentication/providers
Also added a route guard on login and am managing login creds in global state
