+++ package.json
@@ -24,6 +24,11 @@
     "@angular/platform-server": "^19.0.0",
     "@angular/router": "^19.0.0",
     "@angular/ssr": "^19.0.5",
+    "@ngrx/effects": "^19.0.0",
+    "@ngrx/entity": "^19.0.0",
+    "@ngrx/router-store": "^19.0.0",
+    "@ngrx/store": "^19.0.0",
+    "@ngrx/store-devtools": "^19.0.0",
     "express": "^4.18.2",
     "firebase": "^11.1.0",
     "rxjs": "~7.8.0",
