+++ src/app/auth-guard.guard.ts
@@ -1,14 +1,23 @@
 import { inject } from '@angular/core';
+import { Auth } from '@angular/fire/auth';
 import { CanActivateFn, Router } from '@angular/router';
+import { Store } from '@ngrx/store';
+import { AuthState } from './store/reducers/auth.reducer';
+import { map, take } from 'rxjs';
+import { isAuthenticatedSelector } from './store/selectors/auth.selectors';
 
 export const authGuardGuard: CanActivateFn = (route, state) => {
   const router = inject(Router);
-  const isAuthenticated = false; // Replace with your actual authentication check
+  const store = inject(Store<AuthState>);
 
-  if (!isAuthenticated) {
-    router.navigate(['/login']);
-    return false;
-  }
-
-  return true;
+  return store.select(isAuthenticatedSelector).pipe(
+    take(1),
+    map((isAuthenticated) => {
+      if (!isAuthenticated) {
+        router.navigate(['/login']);
+        return false;
+      }
+      return true;
+    })
+  );
 };
