+++ src/app/store/actions/auth.actions.ts
@@ -1,0 +1,4 @@
+import { createAction, props } from '@ngrx/store';
+
+export const login = createAction('[Auth] Login', props<{ userInfo: any }>());
+export const logout = createAction('[Auth] Logout');
